+++ es6/tests/base.js
@@ -96,7 +96,7 @@ describe("Api versioning", function() {
 				name: "APIVersionError",
 				properties: {
 					id: "api_version_error",
-					currentModuleApiVersion: [3, 11, 0],
+					currentModuleApiVersion: [3, 12, 0],
 					neededVersion: [5, 6, 0],
 				},
 			}
@@ -111,7 +111,7 @@ describe("Api versioning", function() {
 				name: "APIVersionError",
 				properties: {
 					id: "api_version_error",
-					currentModuleApiVersion: [3, 11, 0],
+					currentModuleApiVersion: [3, 12, 0],
 					neededVersion: [3, 44, 0],
 				},
 			}